requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
websockets>=11.0.0
supabase>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jwt>=1.3.1
python-multipart>=0.0.6
email-validator>=2.0.0
random-word>=1.0.11
psycopg2>=2.9.5